start of storage system for drill options
more changes on drill options, now hooked up to distributed cache for storing global options.

added test, need to figure out correct response to send to client, and if we want to define a physical operator for setting options

working on adding a test, realized I need to make the distributed map parse incoming values differently by asking the DrillOptions object for what type a given option name is linked to. Thinking of substituting a different implementation of the map in, but for the LocalCache and HazelCache based implementations I would have to create different subclasses of the existing ones.