Drill 342/238 - rebased changes on top of current master, commit 08a9a90d9f.
Removed unneeded duplicate files and updated ignored tests so they pass again. It looks like most of this patch made it in during the large merge, but it never was represented as an independent commit. The changes here are still useful for further consolidating the tests and bringing everything up to date with new interfaces.