WIP dictionary is working.
There is some data that appears corrupted in the dataset used for the testDictionaryError test. This does not appear to be a problem with out reader as a quick comparison of some of the corrupted data read with the parquet-mr cat tool shows the same results. This may have been corrupted by the Pig conversion.