removing some cruft from the top level column refer used by the fixed length fields (will allow for easier re-use of the code for repeated fields)
removed reference to higher level reading method of the scalar data reader from the repeated reader, running into issues trying to read page info twice (as we need to read the page to get the repetition and definition information from the repeated class first to know how much data to read). By invoking the general 'read x values' code in the scalar readers it was trying to read the page again. We will need to have a different model for sharing state information between them.

new test data for repeated.