Support fixed binary reading, for now just reading into var binary vectors as they have functions written against them throughout drill. Still need to add the fixed binary type to drill itself.
WIP - replacing redundant readers for the nullable fixed length types with a single nullable reader that wraps the standard fixed length non-nullable readers. This will remove code redundancy and get us closer to a codebase that can easily be transitioned to code generation.

This also includes implementations for the fixed length non-nullable dictionary readers for ints, floats and double. Previously these were unsupported due to an oversight. As testing was started after completing the non-nullable long dictionary reader, none of the current methods for producing files would make non-nullable files, so the nullable dictionary readers were written and tested. This issue is resolved with the new structure that wraps non-nullable readers with a generic class for angling nullability in a fixed length setting.

Latest issue is in TestJdbcDistQuery, it is referring to the wrong level of a path, I can modify it to work on my machine, but I don't want to break the build elsewhere. This seems to be unrelated to my changes so I assume its some configuration problem with my machine.

Fixed regression in fixed width dictionary readers.

WIP - Still working on a fix for an outstanding issue that looks like it appears at the very end of the read.

Also added support for other fixed dictionary types, int, float, double.

DRILL-1314: impala produced parquet file reading error.

Add ignores back to tests to allow the build to pass for everyone else.