WIP, testing a few outstanding JSON bugs and trying to fix project pushdown.
WIP project pushdown into JSON

WIP - JSON reader enhancements, all text mode has been added, need to figure out how to hook it up to the options system.

Did some testing, we do support scientific notation and currently read it as all of the values with decimal points, as Doubles.

Outstanding request for comment on the list for how to handle int to decimal 'schema changes'.

WIP - more robust test for project pushdown. Need to pull out the part of this branch that is for all text mode, which should be submitted as a separate patch.