WIP - figured out the real problem, need to talk about the best way to fix it.
Currently we consider a column filled with nulls to be the same as a nonexistent one. This has a few disadvantages. Firstly it interacts with project pushdown oddly, we need to figure out how to handle this. This is failing one of the tests.

More importantly it makes the behavior of do nothing when handling nulls dangerous, as we can have a state where we do not have space to store the null. This was the case here, the write into the vector should have failed at index 1000 (as we pre-allocated space for 1000 offsets in a varchar vector). As null handling was a no op we didn't get this message and instead failed improperly on the next index.

This is a bit annoying to fix here. To get the writer for the field I need to know the type unfortunately, this is currently being handled by a switch on the jackson token. We can store a map of the types of known columns, which would allow us to do better error messages without a try/catch around a large section of the code. This still requires that I write new code to populate the holders to send them into the existing setSafe method or we can implement a new setNullSafe method, which seems like the better overall solution. If we declare a new interface implemented by each of the nullable vectors this would give us one method to call for handling this case if we know we have one of the nullable vectors. Otherwise this would require codeine like the rest of the calls to setSafe that take custom holders.

Possibly could look at returning the allocation size from the value vector allocateSafe method, as this changes for each vector based on its previous fill level we do not know how much gets allocated. Getting a max of all of the allocations would give us a record count to stop any operation at.