Cleaning up code in parquet result listener.
Mostly fixed issue with nulls, appears to be one last problem with batch boundaries, but the values are very close to matching up completely.

simple file jumping between runs of 10,000 defined or null values is passing validation against steven's reader. Web returns file from Aman is correct up to index 459

most tests are passing, need to fix one test file still.

More memory and CPU efficient means of checking equality of output of two queries (significantly reduces comparison time comparing against steven's reader)
Reading one of the the files currently still exceeds memory to read all of the data into materialized records, many work on a comparison mechanism that just iterates through the batches together (they will not line up batch for batch, so we need to concurrent iterators)

There are two nullable files failing 958 and the new generated file with 10,000 values then 10,000 nulls back and forth for 2 million records. It is only for the one that starts with values. The same pattern starting with a run of nulls works.

the test with 10,000 rows switching between all nulls and all defined values that started with nulls only seemed to work because of a poor choice of input. Unique values used for non-null values revealed an issue, now fixed. The one starting with defined values is hanging.

Updated test to match new file, previously I had filled the file with 1's so I would know the sum, but then I didn't guarantee that I had read the right data into the right places within the vectors.

WIP - was experiencing some issues with data being transferred into the vector, but the nullability at some positions was off. Tried adding tracking for the number of definition levels read (for the current vector, this should be equal to the position we want to set the nullability for at any time)

Still not working correctly, just wanted to save this in the history as I continue

Both of the tests switching between 10,000 null or defined values pass, as well as several others. But now some that were working before are broken.

everything but the 958 test passes

Everything except the file for Drill 1314 is reading correctly, there just seems to be an issue late in the read with count across vectors.

With these changes I think there are several bits of cruft that made there way into the reader the numerous times I was debugging issues like this, I believe the current implementation is using a simplified approach, but I have to go back and try to remove some stuff and review correctness.

Everything is working! To test the SF1 supplier file reading with both of the readers I had to disable parallelization. It didn't seem to respect the max width per node setting, so I went in and modified the SimpleParalellizer directly. I backed out the changes after the test passed.

Still need to do a little cleanup. The last real issue fixed was using the Uncompressed size of the column chunk to set up the ColumnDataReader, rather than the compressed size, this was causing mismatches in the batch sizes of some of the columns at the very end of the read.

Re-enable fixed binary reading that was mistakenly taken out of the execution path while reverting some issues before the 0.5 release candidate.