Everything is working! To test the SF1 supplier file reading with both of the readers I had to disable parallelization. It didn't seem to respect the max width per node setting, so I went in and modified the SimpleParalellizer directly. I backed out the changes after the test passed.
Still need to do a little cleanup. The last real issue fixed was using the Uncompressed size of the column chunk to set up the ColumnDataReader, rather than the compressed size, this was causing mismatches in the batch sizes of some of the columns at the very end of the read.