Single column varchar is matching between the two parquet readers with the new faster layout that matches parquet format.
This needs to be enhanced to use a start/end rather than just a constant buffer between each value, the issue is that no data (including a length) is given in the parquet file if the data is null. This requires us to specify a start and end.

Conflicts:
	exec/java-exec/src/main/java/org/apache/drill/exec/planner/fragment/SimpleParallelizer.java
	exec/java-exec/src/main/java/org/apache/drill/exec/store/parquet/columnreaders/NullableVarLengthValuesColumn.java

Conflicts:
	exec/java-exec/src/test/java/org/apache/drill/exec/physical/impl/writer/TestParquetWriter.java