Fix for the failed Regression test run. Was an issue with removing cast information, simply needed to return the original expression rather than try to clone it if rewriting was not needed.
failing tpch 5 on clust, sf-10 works locally

WIP fixing repeated list

WIP - TODO - current rule for changing convert_from(expr, 'json') into the convert_fromJSON function does not work if the expression is nested down in the expression tree. This should work with the new complex expression splitter rule, but this rewrite rule needs to re-write the expression before it can be determined to have a complex output, and this currently only happens at the root expression level (as we previously did not support complex expressions deeper down in the expression tree).

Was not visiting the child of a project appropriately, this was causing complex expressions nested inside of a sub-query to fail.

go back to returning a clone of the call in the complex expresso splitting rex visitor.