Drill-1688: Complex parquet reader fails to read wide records.
Fixed complex parquet reader to allow configuring the threshold at which the vectors being populated are checked for fill level, this is the means of mitigating the parquet-mrs inability to re-wind a read operation of one of the writes fails. The issue is the threshold needs to be adjustable to allow a balance between performance and allowing wider records.