WIP - small change to the fix, as well as a secondary change that disables the fix to standardize the behavior with the rest of drill.
If I call count on a repeated field read from json it gives me an error about function not found. With the change I made trying to make the behavior consistent, I see this error in the log, but I actually get another error about setSchema not being called. Should probably open a jira to improve the behavior of function resolution if we try to call a udf on a type that it is not defined on (as we often read out of files with a type ANY, and we have to assume that the type that arrives in the udf will be correct)