DRILL-2173: Partition queries to drive dynamic pruning
Adds new interface on storage plugin registry as well as individual storage plugins for exploring workspaces.
Adds injection for partition explorer to expression materialization and interpreted evaluation.

New interface on storage plugin registry as well as individual storage plugins for exploring workspaces.

Tasks remaining:
- code cleanup
- review if calcite rule fork is necessary
- should at least put in in the drill fork
- get timezone and query start time available to query context, currently created while running physical plan
- prevents folding calls to now()
- finish the switch statement in DrillConstExecutor to allow folding all drill types into literals

Conflicts:
	exec/java-exec/src/main/java/org/apache/drill/exec/ops/QueryContext.java

2173 update -Address Chris' review comments.

Change the PartitionExplorer to return an Iterable<String> instead of String[]

Add interface level description to PartitionExplorer and StoragePluginPartitionExplorer.

New inner class in FileSystemPlugin to fulfill the new Iterable interface for partitions.

Formatting/cleanup fixes

Clean up error reporting code in MaxDir UDF. Remove method to get a string from a DrillBuf, as it was already defined in StringFunctionHelpers. Add new utility method to specifically convert a VarCharHolder to a string to remove some boilerplate.

Fixed an errant copy paste in a comment and removed unused imports.

Fix docs in FileSystemPlugin, belongs with the 2173 changes.

Fix references in Javadoc to properly use @link instead of @see.

2173 fixes, correctly return an empty list of sub-partitions if the path requested in the partition explorer interface is a file. Fix a few docs.

More 2173, finishing Chris' comments

2173 update - Add validation for PartitionExplorer injectable in UdfUtiltiers.

Conflicts:
	exec/java-exec/src/main/java/org/apache/drill/exec/ops/FragmentContext.java

small change to fix refactored unit tests.

cleanup for 2173

Fix maxdir UDF so it can compile in runtime generated code as well as the interpreted expression system (needed to fully qualify classes and interfaces). It still fails to execute, as we prevent requesting a schema from a non-root fragment. We do not expect these types of functions to ever be used without constant folding so this should not be an issue.