WIP - testing intermittent TPCH distributed failures
Several test in TestAgggateFunctions are failing with planning issues even when I disable the new rules.

They do not fail if I run them without 2092 test, this succeeds, but causes the subsequent tests to fail.

WIP

TODO - look back at this

Random failures in tech distributed complaining about child allocator inside of fragment context not closed. Disabling to see the status of the rest of the tests.

WIP debugging