WIP
WIP

WIP

WIP

WIP

Fix folding expressions that result in null after refactoring the interpreted expression evaluation to return a ValueHolder in the case of a constant expression (instead of a value vector which calling getObject on previously gracefully handled nulls as they were passed into the Calciate API to create literals).

One last fix, was just checking nullability of output value, not the actual flag for null (which is different for each holder, but there was already a utility method with the switch statement).

WIP