Reduce the default fork count
The build has been having trouble in several devs Linux VMs, previously we would advise turning down the fork count, which used to be hard coded to 4.

It was changed to be proportionate to number of cores, with 1 fork for each core. This setting has been producing timeouts as there seems to be some
contetion between the forks, or there are some tests that are too close to the timeout.