Finished manually re-applying rejected hunks from Jacques patch.
WIP - moving tests to remove references to now deleted option name constants

Confirmed behavior is correct, changing TODO into a descriptive comment

Fixing up one more reference to an old string constant for an option name in the jdbc tests.

Update most of the unit tests to use the new interface for setting option values, fix a few last places in the code code that were not using the new interfaces for getting the values out of the option manager.

fix a few more references to old option name constants

adding missing finals

Move the implementations of the methods to set and reset option values outside of BaseTestQuery so they can be shared with test classes that do not inherit from it. Use this refactoring in TestDrillbitResilience and fix a few option setting/resetting calls in TestExtendedTypes.

WIP

Combine OptionValidator with its one direct subclass TypeValidator, removed a little cruft and allows for simple usage of the type expected by an option in an exception message.

Small fix for the getDefaultString method on consolidated OptionValidator class hierarchy

WIP - writing a unit test with verification for the sys.options table

Fixed issue with the drill bit control option being defined as a subclass of a DoubleValidator instead of StringValidator. Finished test with validation on the options table as well as a basic test for setting options at the system and session level.

Got the test with simple root exec working, need to update all the tests like this or figure out what in this patch broke these tests.

Revert non-standard option access in ScanBatch to allow SimpleRootExec tests to keep running.

Reset was failing for String options because the values were not enclosed in single quotes.

WIP - move this back to system level settings, make a reference to similar test in TestOptions

reset system option after test

Revert changes to TestCastEmprtStrings

Better fix

Fixing more tests