WIP- 2838 and question about failing flatten from list
WIP - copied code from Project to fix the broken case, would really like to find a way to just make the code common between them.

Test case for large list flatten, issue is actually with the fact that the flatten planning is copying the whole list to be flattened next to each outgoing record. For large lists this is causing a memory vallooning effect that is killing queries.

WIP - trying to enable previously disabled test on flatten

Separate test cases for failure in flatten vs failure in json scan with large lists.