Fixed the last failing parquet test.
The issue was a mismatch in the usage of byte buffers. Even though the position of a buffer was being set, that seemed to be ignored in the setSafe method on the varbinary vector. I needed to pass in the offset as it seems to just read from the beginning of the buffer. I'm not sure this is how ByteBuffers are supposed to be used, but we seem to make use of this pattern commonly so I'm not sure it could be easily refactored.

Added some test to print out some additional context when an ordered comparison of two datasets fails in a test.