Add tests for user disabled corrupt date correction
In some cases we cannot absolutely determine if the dates values are
corrupt based on the metadata. This option disables corrupt data correction,
only in the case where we aren't sure if Drill wrote it. If we know from
the metadata that Drill wrote the file, corrupt or correct, the behavior
of the reader will not change.

Move tests to their own class.