Add test reading old metadata cache file containing corrupt dates
Requires a small hack to insert the full path into the file. I want this
generated with an old version of Drill, but all of the paths are absolute,
so I cannot generate it on my machine with a path I know will be available
on any future build machine. Hacky solution is to put in a string to
substitute in the test with a path once I generate a unique temp directory
and copy the parquet files there.