Temporary fix for build issue with generated dependency-reduced-pom.xml
This generated pom file was being discovered and maven was trying to
run the target directory in jdbc-all as a submodule.

This change reverts to the default output location (the module root)
and adds corresponding .gitignore and RAT exclude entries.

More investigation of why this became an issue when we added the
maven-enforcer plugin to the module (and only appears when running a release)
will come in DRILL-4336.