DRILL-4445: Standardize the Physical and Logical plan nodes to use Lists instead of arrays for their inputs
Remove some extra translation logic used to move between the
two representations.

TODO - look back the the Join logical node, has two JsonCreator annotations,
but only one will be used. Not sure if the behavior of which is chosen
is considered documented behavior, should just fix it on our end.